Moved template evaluator class from databinds package.
